requirements.txt
# Discord Bot Framework
discord.py>=2.3.0
discord-py-interactions>=1.0.0
discord-typings>=0.9.0
PyNaCl>=1.5.0

# API et Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
blagues_api>=1.0.0

# Utilitaires
emoji>=2.8.0
youtube-dl>=2021.12.17
croniter>=1.4.0
setuptools>=68.0.0
bs4>=0.0.1
Pillow>=10.0.0

# Interface et Logging
rich>=13.0.0
python-dotenv>=1.0.0

# Base de données PostgreSQL
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Tests et développement
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0 